Rename internal/io to internal/fsutil
This avoids import collisions with go's io package
