+++ cache/volume_cache.go
@@ -2,7 +2,7 @@ package cache
 
 import (
 	"encoding/json"
-	goio "io"
+	"io"
 	"os"
 	"path/filepath"
 	"runtime"
@@ -10,7 +10,7 @@ import (
 
 	"github.com/pkg/errors"
 
-	"github.com/buildpacks/lifecycle/internal/io"
+	"github.com/buildpacks/lifecycle/internal/fsutil"
 	"github.com/buildpacks/lifecycle/platform"
 )
 
@@ -106,13 +106,13 @@ func (c *VolumeCache) AddLayerFile(tarPath string, diffID string) error {
 		return nil
 	}
 
-	if err := io.Copy(tarPath, layerTar); err != nil {
+	if err := fsutil.Copy(tarPath, layerTar); err != nil {
 		return errors.Wrapf(err, "caching layer (%s)", diffID)
 	}
 	return nil
 }
 
-func (c *VolumeCache) AddLayer(rc goio.ReadCloser, diffID string) error {
+func (c *VolumeCache) AddLayer(rc io.ReadCloser, diffID string) error {
 	if c.committed {
 		return errCacheCommitted
 	}
@@ -123,7 +123,7 @@ func (c *VolumeCache) AddLayer(rc goio.ReadCloser, diffID string) error {
 	}
 	defer fh.Close()
 
-	if _, err := goio.Copy(fh, rc); err != nil {
+	if _, err := io.Copy(fh, rc); err != nil {
 		return errors.Wrap(err, "copying layer to tar file")
 	}
 	return nil
@@ -139,7 +139,7 @@ func (c *VolumeCache) ReuseLayer(diffID string) error {
 	return nil
 }
 
-func (c *VolumeCache) RetrieveLayer(diffID string) (goio.ReadCloser, error) {
+func (c *VolumeCache) RetrieveLayer(diffID string) (io.ReadCloser, error) {
 	path, err := c.RetrieveLayerFile(diffID)
 	if err != nil {
 		return nil, err
@@ -177,13 +177,13 @@ func (c *VolumeCache) Commit() error {
 		return errCacheCommitted
 	}
 	c.committed = true
-	if err := io.RenameWithWindowsFallback(c.committedDir, c.backupDir); err != nil {
+	if err := fsutil.RenameWithWindowsFallback(c.committedDir, c.backupDir); err != nil {
 		return errors.Wrap(err, "backing up cache")
 	}
 	defer os.RemoveAll(c.backupDir)
 
-	if err1 := io.RenameWithWindowsFallback(c.stagingDir, c.committedDir); err1 != nil {
-		if err2 := io.RenameWithWindowsFallback(c.backupDir, c.committedDir); err2 != nil {
+	if err1 := fsutil.RenameWithWindowsFallback(c.stagingDir, c.committedDir); err1 != nil {
+		if err2 := fsutil.RenameWithWindowsFallback(c.backupDir, c.committedDir); err2 != nil {
 			return errors.Wrap(err2, "rolling back cache")
 		}
 		return errors.Wrap(err1, "committing cache")
