+++ internal/fsutil/testdata/some_dir/other_dir/other_link
