+++ internal/fsutil/testdata/some_dir/some_link
